Created Input File
Gets the users input as a string from the keyboard and makes sure that the character count is 4. It also catches values that are not valid and requests the user to insert the data again. The function then reprints the data once its done to the user, so they can validate that its correct